refactor: StationPath 삭제 및 Sections로 이동
