+++ src/main/java/wooteco/subway/domain/Sections.java
@@ -5,9 +5,12 @@ import static java.util.stream.Collectors.toList;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import wooteco.subway.dto.StationResponse;
+import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
+import org.jgrapht.graph.DefaultWeightedEdge;
+import org.jgrapht.graph.WeightedMultigraph;
 import wooteco.subway.utils.exception.SectionCreateException;
 import wooteco.subway.utils.exception.SectionDeleteException;
+import wooteco.subway.utils.exception.StationNotFoundException;
 import wooteco.subway.utils.exception.SubwayException;
 
 public class Sections {
@@ -18,6 +21,11 @@ public class Sections {
     private static final String SECTION_NOT_CONNECT_MESSAGE = "구간이 연결되지 않습니다";
     private static final String SECTION_MUST_SHORTER_MESSAGE = "기존의 구간보다 긴 구간은 넣을 수 없습니다.";
     private static final int MIN_SIZE = 1;
+    private static final int DEFAULT_DISTANCE = 10;
+    private static final int DEFAULT_FARE = 1250;
+    private static final int OVER_FARE_DISTANCE = 50;
+    private static final int STANDARD_UNIT = 5;
+    private static final int MAX_UNIT = 8;
 
     private final List<Section> values;
 
@@ -177,7 +185,54 @@ public class Sections {
                 .findFirst();
     }
 
-    public boolean isExistStation(final Station station) {
+    public int calculateMinDistance(final Station startStation, final Station endStation) {
+        validateExistStation(startStation, endStation);
+        return (int) createSectionDijkstraShortestPath().getPathWeight(startStation, endStation);
+    }
+
+    private DijkstraShortestPath<Station, DefaultWeightedEdge> createSectionDijkstraShortestPath() {
+        WeightedMultigraph<Station, DefaultWeightedEdge> graph
+                = new WeightedMultigraph<>(DefaultWeightedEdge.class);
+        for (Station station : sortSections()) {
+            graph.addVertex(station);
+        }
+        for (Section section : values) {
+            assignWeight(graph, section);
+        }
+        return new DijkstraShortestPath<>(graph);
+    }
+
+    private void assignWeight(final WeightedMultigraph<Station, DefaultWeightedEdge> graph, final Section section) {
+        graph.setEdgeWeight(graph.addEdge(section.getUpStation(), section.getDownStation()), section.getDistance());
+    }
+
+    private void validateExistStation(Station startStation, Station endStation) {
+        if (!isExistStation(startStation) || !isExistStation(endStation)) {
+            throw new StationNotFoundException();
+        }
+    }
+
+    public List<Station> findShortestStations(final Station startStation, final Station endStation) {
+        return createSectionDijkstraShortestPath().getPath(startStation, endStation).getVertexList();
+    }
+
+    public int calculateFare(final int distance) {
+        if (distance <= DEFAULT_DISTANCE) {
+            return DEFAULT_FARE;
+        }
+        if (distance <= OVER_FARE_DISTANCE) {
+            return DEFAULT_FARE + calculateOverFare(distance - DEFAULT_DISTANCE, STANDARD_UNIT);
+        }
+        return DEFAULT_FARE
+                + calculateOverFare(OVER_FARE_DISTANCE - DEFAULT_DISTANCE, STANDARD_UNIT)
+                + calculateOverFare(distance - OVER_FARE_DISTANCE, MAX_UNIT);
+    }
+
+    private int calculateOverFare(final int distance, final int unit) {
+        return (int) ((Math.ceil((distance - 1) / unit) + 1) * 100);
+    }
+
+    private boolean isExistStation(final Station station) {
         return values.stream()
                 .anyMatch(section -> section.have(station));
     }
